refactor out omnipanelviewindex and file path changes
