+++ packages/dashboard/src/components/dashboard/dashboard-config.tsx
@@ -1,0 +1,10 @@
+export enum OmniPanelViewIndex {
+  MainMenu = 0,
+  Doors,
+  Lifts,
+  Robots,
+  Dispensers,
+  Commands,
+  Negotiations,
+  Tasks,
+}
