+++ packages/dashboard/src/components/dashboard/reducers/dashboard-reducer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { StackNavigator } from 'react-components';
-import { OmniPanelViewIndex } from '../dashboard';
+import { OmniPanelViewIndex } from '../dashboard-config';
 
 type DashboardActionFormat<T, K = void> = K extends void
   ? {
