+++ packages/dashboard/src/components/dashboard/tour/tour.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Tour from 'reactour';
-import { OmniPanelViewIndex } from '../dashboard';
+import { OmniPanelViewIndex } from '../dashboard-config';
 import { createTourSteps } from './tour-manager';
 
 export interface DashboardTourProps {
