+++ packages/dashboard/src/hotkeys.ts
@@ -1,6 +1,6 @@
 import { HotKeysEnabledProps, KeyMap, KeySequence } from 'react-hotkeys';
 import { AppController } from './components/app-contexts';
-import { OmniPanelViewIndex } from './components/dashboard/dashboard';
+import { OmniPanelViewIndex } from './components/dashboard/dashboard-config';
 import { ReducerDashboardDispatch } from './components/dashboard/reducers/dashboard-reducer';
 
 export interface hotKeysProps {
