+++ packages/dashboard/src/tests/hotkeys.test.ts
@@ -1,7 +1,8 @@
 import { act, HookResult, renderHook } from '@testing-library/react-hooks';
 import { HotKeysEnabledProps } from 'react-hotkeys';
 import { AppController } from '../components/app-contexts';
-import { dashboardInitialValues, OmniPanelViewIndex } from '../components/dashboard/dashboard';
+import { dashboardInitialValues } from '../components/dashboard/dashboard';
+import { OmniPanelViewIndex } from '../components/dashboard/dashboard-config';
 import {
   ReducerDashboardProps,
   useDashboardReducer,
